--- a/packages/cli/src/cli.ts
+++ b/packages/cli/src/cli.ts
@@ -107,7 +107,7 @@ yargs(process.argv.slice(2))
           } else {
             console.warn(`No ${field} in ${appMapBaseName}`);
           }
-        }, 5);
+        }, 2);
         appMapNames.forEach((name) => q.push(name));
         if (!q.idle()) await q.drain();
       } else {