--- a/packages/cli/src/cli.ts
+++ b/packages/cli/src/cli.ts
@@ -109,7 +109,7 @@ yargs(process.argv.slice(2))
           }
         }, 5);
         appMapNames.forEach((name) => q.push(name));
-        await q.drain();
+        if (!q.idle()) await q.drain();
       } else {
         appMapNames.forEach((name) => values.push(name));
       }