* Optimize target file enumeration
* Cache file modified times
* Remove the most verbose logging
* Remove source locations that are blank

Before optimization:
```
node ./built/cli.js depends -d ~/source/land-of-apps/forem > forem.log  11.41s user 6.40s system 169% cpu 10.537 total
```

After optimization:
```
node ./built/cli.js depends -d ~/source/land-of-apps/forem > forem.log  3.86s user 1.59s system 127% cpu 4.296 total
```