@appland-release 2023-05-24
:tada: This PR is included in version @appland/scanner-v1.77.2 :tada:

The release is available on:
- `@appland/scanner-v1.77.2`
- `@appland/scanner-v1.77.2`
- [GitHub release](https://github.com/getappmap/appmap-js/releases/tag/%40appland/scanner-v1.77.2)
- `@appland/scanner-v1.77.2`

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: